the architecture of the project has been finalized
