--- Layout-of-the-iPhone-settings-screen-using-UITableView/Layout-of-the-iPhone-settings-screen-using-UITableView/AppDelegate.swift
@@ -1,46 +1,0 @@
-//
-//  AppDelegate.swift
-//  Layout-of-the-iPhone-settings-screen-using-UITableView
-//
-//  Created by Georgiy on 02.06.2022.
-//
-
-import UIKit
-
-@main
-class AppDelegate: UIResponder, UIApplicationDelegate {
-
-
-
-    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
-        if #available(iOS 15, *) {
-              let navigationBarAppearance = UINavigationBarAppearance()
-              navigationBarAppearance.configureWithDefaultBackground()
-              UINavigationBar.appearance().standardAppearance = navigationBarAppearance
-              UINavigationBar.appearance().scrollEdgeAppearance = navigationBarAppearance
-
-              let tabBarAppearance = UITabBarAppearance()
-              tabBarAppearance.configureWithDefaultBackground()
-              UITabBar.appearance().standardAppearance = tabBarAppearance
-              UITabBar.appearance().scrollEdgeAppearance = tabBarAppearance
-            }
-        return true
-    }
-
-    // MARK: UISceneSession Lifecycle
-
-    func application(_ application: UIApplication, configurationForConnecting connectingSceneSession: UISceneSession, options: UIScene.ConnectionOptions) -> UISceneConfiguration {
-        // Called when a new scene session is being created.
-        // Use this method to select a configuration to create the new scene with.
-        return UISceneConfiguration(name: "Default Configuration", sessionRole: connectingSceneSession.role)
-    }
-
-    func application(_ application: UIApplication, didDiscardSceneSessions sceneSessions: Set<UISceneSession>) {
-        // Called when the user discards a scene session.
-        // If any sessions were discarded while the application was not running, this will be called shortly after application:didFinishLaunchingWithOptions.
-        // Use this method to release any resources that were specific to the discarded scenes, as they will not return.
-    }
-
-
-}
-
