+++ Layout-of-the-iPhone-settings-screen-using-UITableView/Layout-of-the-iPhone-settings-screen-using-UITableView/Resources/Assets.xcassets/AccentColor.colorset/Contents.json
