+++ Layout-of-the-iPhone-settings-screen-using-UITableView/Layout-of-the-iPhone-settings-screen-using-UITableView/Resources/Assets.xcassets/vpn.imageset/Contents.json
