+++ Layout-of-the-iPhone-settings-screen-using-UITableView/Layout-of-the-iPhone-settings-screen-using-UITableView/Sources/App/AppDelegate.swift
@@ -1,0 +1,17 @@
+//
+//  AppDelegate.swift
+//  Layout-of-the-iPhone-settings-screen-using-UITableView
+//
+//  Created by Georgiy on 02.06.2022.
+//
+
+import UIKit
+
+@main
+class AppDelegate: UIResponder, UIApplicationDelegate {
+
+    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
+        
+        return true
+    }
+}
