+++ Layout-of-the-iPhone-settings-screen-using-UITableView/Layout-of-the-iPhone-settings-screen-using-UITableView/Sources/Scenes/Model/ExtensionSettingsModel.swift
