+++ Layout-of-the-iPhone-settings-screen-using-UITableView/Layout-of-the-iPhone-settings-screen-using-UITableView/Sources/Scenes/View/SettingTableFigure.swift
